+++ src/main/java/com/CabbageAndGarlic/repository/RoutingRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RoutingRepository extends JpaRepository<Routing, Long> {
-    List<Routing> findByNumber(Long routingNumber);
+public interface RoutingRepository extends JpaRepository<Routing, Integer> {
+    List<Routing> findByNumber(int routingNumber);
 }
